repository - 공통 repository외 모듈별 분리 security - java17에맞게 변경 모듈 - admin,user api 구분

+++ admin/src/main/java/com/example/module/api/member/service/MemberService.java
@@ -1,12 +1,12 @@
 package com.example.module.api.member.service;
 
 import com.example.module.api.member.dto.request.RequestMemberDto;
+import com.example.module.api.member.dto.request.RequestPatchMemberDto;
 import com.example.module.api.member.dto.response.ResponseMemberDto;
 import com.example.module.entity.Member;
 import com.example.module.repository.TokenRepository;
 import com.example.module.repository.member.MemberRepository;
 import com.example.module.util.CommonException;
-import com.example.module.util.SecurityContextHelper;
 import com.example.module.util._Enum.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -42,17 +42,21 @@ public class MemberService {
                         .sex(memberCreateDto.getSex())
                         .age(memberCreateDto.getAge())
                         .phoneNumber(memberCreateDto.getPhoneNumber())
-                        .roles(List.of("USER"))
+                        .roles(List.of(memberCreateDto.getRole()))
                         .build());
 
     }
 
     @Transactional
     public void deleteMember(Member member) {
-        // 관리자/일반유저 구분
-        SecurityContextHelper.isAuthorizedForMember(member);
+        if(member.getRoles().contains("ADMIN")){
+            throw new CommonException(ErrorCode.ACCESS_DENIED);
+        }
 
         tokenRepository.deleteByTokenKey(member.getUserId());
+
+        member.setDeleted(true);
+        memberRepository.save(member);
     }
 
     public Page<ResponseMemberDto> getMemberList(HashMap<String, Object> filters, Pageable pageable) {
@@ -63,6 +67,13 @@ public class MemberService {
         if(memberRepository.findByUserId(id).isPresent()){
             throw new CommonException(ErrorCode.MEMBER_DUPLICATED);
         }
+    }
+
+    @Transactional
+    public void patchMember(Member member,RequestPatchMemberDto requestPatchMemberDto) {
+        // 정책에 따라 권한이 바뀔수잇지만 관리자가
+        member.setRoles(List.of(requestPatchMemberDto.getRole()));
+        memberRepository.save(member);
     }
 
 //    private String resolveToken(String accessToken) {
